Refactor and redfine what an arena is
