--- src/main/java/io/github/minigamecore/api/arena/data/types/IntegerType.java
@@ -1,52 +1,0 @@
-/*
- * This file is part of MinigameCoreAPI, licensed under the MIT License (MIT).
- *
- * Copyright (c) 2016 - 2016 MinigameCore <http://minigamecore.github.io>
- * Copyright (c) Contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package io.github.minigamecore.api.arena.data.types;
-
-import io.github.minigamecore.api.arena.data.DataType;
-
-public class IntegerType implements DataType<Integer> {
-
-    @Override
-    public String getId() {
-        return "minigamecore:integer";
-    }
-
-    @Override
-    public String getName() {
-        return "Integer";
-    }
-
-    @Override
-    public Integer deserialize(String input) {
-        return Integer.valueOf(input);
-    }
-
-    @Override
-    public String serialize(Integer input) {
-        return input.toString();
-    }
-
-}
