+++ src/main/java/io/github/minigamecore/api/data/types/LocationType.java
@@ -23,10 +23,11 @@
  * THE SOFTWARE.
  */
 
-package io.github.minigamecore.api.arena.data.types;
+package io.github.minigamecore.api.data.types;
 
-import io.github.minigamecore.api.arena.data.DataType;
-import io.github.minigamecore.api.arena.data.InvalidDataException;
+import io.github.minigamecore.api.data.DataType;
+import io.github.minigamecore.api.data.InvalidDataException;
+
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.world.Location;
 import org.spongepowered.api.world.World;
