+++ src/main/java/io/github/minigamecore/api/event/ArenaEvent.java
@@ -25,7 +25,8 @@
 
 package io.github.minigamecore.api.event;
 
-import io.github.minigamecore.api.arena.Arena;
+import io.github.minigamecore.api.minigame.arena.Arena;
+
 import org.spongepowered.api.event.Event;
 
 /**
